+++ jest.config.js
@@ -1,4 +1,17 @@
 module.exports = {
+  coveragePathIgnorePatterns: [
+    "/node_modules/",
+    "<rootDir>/src/test-utilities.ts",
+  ],
+  coverageThreshold: {
+    "global": {
+      "branches": 90,
+      "functions": 90,
+      "lines": 90,
+      "statements": 90,
+    }
+  },
+  mapCoverage: true,
   moduleFileExtensions: [
     'ts', 'js',
   ],