+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- **BREAKING** Fix type of `subreducer` to use 4 type parameters.  
+  It used originally 3 type parameters.
+    - Following type is fixed.
+        ```typescript
+        subreducer(action('a', payload<number>()), (...args: any): any => ({}));
+        // Type before fixed: SingleActionReducer<any, 'a', any>
+        // Type after fixed: SingleActionReducer<any, 'a', number>
+        ```
+
 ## [1.1.0] - 2018-02-17
 
 ### Added