+++ CHANGELOG.md
@@ -36,8 +36,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Miscellaneous
 - Add tests
-    - positive/negative tests for action.
-    - positive/negative tests for action type.
+    - positive/negative tests for action functions.
+    - positive/negative tests for action type functions.
 
 ## [1.0.1] - 2018-02-10