+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Miscellaneous
 - Add tests
     - positive/negative tests for reducer functions.
+- Add travis supports
 
 ## [1.1.0] - 2018-02-17