+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add tests
     - positive/negative tests for reducer functions.
 - Add travis supports
+- Add codecov supports
 
 ## [1.1.0] - 2018-02-17